+++ app/mailers/alaveteli_pro/embargo_mailer.rb
@@ -75,8 +75,7 @@ module AlaveteliPro
         site_name: site_name.html_safe,
         count: info_requests.count
       )
-      auto_generated_headers
-      mail_user(@user, subject)
+      mail_user(@user, subject: subject)
     end
 
     def expired_alert(user, info_requests)
@@ -89,29 +88,7 @@ module AlaveteliPro
         site_name: site_name.html_safe,
         count: info_requests.count
       )
-      auto_generated_headers
-      mail_user(@user, subject)
-    end
-
-    private
-
-    # TODO: these are copied from request_mailer, but it seems like they should
-    # be something shared via application_mailer.
-    def auto_generated_headers
-      headers({
-        'Return-Path' => blackhole_email,
-        'Reply-To' => pro_contact_from_name_and_email, # not much we can do if the user's email is broken
-        'Auto-Submitted' => 'auto-generated', # http://tools.ietf.org/html/rfc3834
-        'X-Auto-Response-Suppress' => 'OOF'
-      })
-    end
-
-    def mail_user(user, subject)
-      mail({
-        from: pro_contact_from_name_and_email,
-        to: user.name_and_email,
-        subject: subject
-      })
+      mail_user(@user, subject: subject)
     end
   end
 end