+++ app/mailers/request_mailer.rb
@@ -69,7 +69,7 @@ class RequestMailer < ApplicationMailer
     @info_request = info_request
     @message = message
 
-    set_reply_to_headers(nil, 'Reply-To' => user.name_and_email)
+    set_reply_to_headers('Reply-To' => user.name_and_email)
 
     # From is an address we control so that strict DMARC senders don't get refused
     mail(from: MailHandler.address_from_name_and_email(
@@ -88,14 +88,10 @@ class RequestMailer < ApplicationMailer
     @incoming_message = incoming_message
     @info_request = info_request
 
-    set_reply_to_headers(info_request.user)
-    set_auto_generated_headers
-
-    mail(
-      from: contact_for_user(info_request.user),
-      to: info_request.user.name_and_email,
+    mail_user(
+      info_request.user,
       subject: _("New response to your FOI request - {{request_title}}",
-                    request_title: info_request.title.html_safe),
+                 request_title: info_request.title.html_safe),
       charset: "UTF-8"
     )
   end
@@ -105,13 +101,10 @@ class RequestMailer < ApplicationMailer
     @url = respond_to_last_url(info_request)
     @info_request = info_request
 
-    set_reply_to_headers(user)
-    set_auto_generated_headers
-
     mail_user(
       user,
-      _("Delayed response to your FOI request - {{request_title}}",
-        request_title: info_request.title.html_safe)
+      subject: _("Delayed response to your FOI request - {{request_title}}",
+                 request_title: info_request.title.html_safe)
     )
   end
 
@@ -120,13 +113,11 @@ class RequestMailer < ApplicationMailer
     @url = respond_to_last_url(info_request)
     @info_request = info_request
 
-    set_reply_to_headers(user)
-    set_auto_generated_headers
-
     mail_user(
       user,
-      _("You're long overdue a response to your FOI request - {{request_title}}",
-        request_title: info_request.title.html_safe)
+      subject: _("You're long overdue a response to your FOI request - " \
+                 "{{request_title}}",
+                 request_title: info_request.title.html_safe)
     )
   end
 
@@ -140,12 +131,10 @@ class RequestMailer < ApplicationMailer
     @incoming_message = incoming_message
     @info_request = info_request
 
-    set_reply_to_headers(info_request.user)
-    set_auto_generated_headers
     mail_user(
       info_request.user,
-      _("Please update the status of your request - {{request_title}}",
-        request_title: info_request.title.html_safe)
+      subject: _("Please update the status of your request - {{request_title}}",
+                 request_title: info_request.title.html_safe)
     )
   end
 
@@ -154,10 +143,10 @@ class RequestMailer < ApplicationMailer
     @url = request_url(info_request)
     @info_request = info_request
 
-    set_reply_to_headers(info_request.user)
-    set_auto_generated_headers
-    mail_user(info_request.user, _("Someone has updated the status of " \
-                                      "your request"))
+    mail_user(
+      info_request.user,
+      subject: _("Someone has updated the status of your request")
+    )
   end
 
   # Tell the requester that they need to clarify their request
@@ -171,12 +160,10 @@ class RequestMailer < ApplicationMailer
     @incoming_message = incoming_message
     @info_request = info_request
 
-    set_reply_to_headers(info_request.user)
-    set_auto_generated_headers
     mail_user(
       info_request.user,
-      _("Clarify your FOI request - {{request_title}}",
-        request_title: info_request.title.html_safe)
+      subject: _("Clarify your FOI request - {{request_title}}",
+                 request_title: info_request.title.html_safe)
     )
   end
 
@@ -186,12 +173,10 @@ class RequestMailer < ApplicationMailer
     @info_request = info_request
     @url = comment_url(comment)
 
-    set_reply_to_headers(info_request.user)
-    set_auto_generated_headers
     mail_user(
       info_request.user,
-      _("Somebody added a note to your FOI request - {{request_title}}",
-        request_title: info_request.title.html_safe)
+      subject: _("Somebody added a note to your FOI request - {{request_title}}",
+                 request_title: info_request.title.html_safe)
     )
   end
 
@@ -202,12 +187,10 @@ class RequestMailer < ApplicationMailer
     @info_request = info_request
     @url = comment_url(earliest_unalerted_comment)
 
-    set_reply_to_headers(info_request.user)
-    set_auto_generated_headers
     mail_user(
       info_request.user,
-      _("Some notes have been added to your FOI request - {{request_title}}",
-        request_title: info_request.title.html_safe)
+      subject: _("Some notes have been added to your FOI request - {{request_title}}",
+                 request_title: info_request.title.html_safe)
     )
   end