+++ spec/mailers/notification_mailer_spec.rb
@@ -374,7 +374,7 @@ RSpec.describe NotificationMailer do
 
     it "send the message from the right address" do
       mail = NotificationMailer.daily_summary(user, all_notifications)
-      expect(mail.from).to eq ['postmaster@localhost']
+      expect(mail.from).to eq [blackhole_email]
     end
 
     it "sets the right subject line" do
@@ -393,11 +393,8 @@ RSpec.describe NotificationMailer do
       expect(mail.body.encoded).to eq(expected_message)
     end
 
-    it "sets reply_to headers" do
+    it "sets mail headers" do
       mail = NotificationMailer.daily_summary(user, all_notifications)
-      expected_reply_to = "#{AlaveteliConfiguration.contact_name} " \
-                          "<#{AlaveteliConfiguration.contact_email}>"
-      expect(mail.header["Reply-To"].value).to eq expected_reply_to
       expect(mail.header["Return-Path"].value).
         to eq 'do-not-reply-to-this-address@localhost'
     end
@@ -465,14 +462,11 @@ RSpec.describe NotificationMailer do
 
     it "sends the message from the right address" do
       mail = NotificationMailer.response_notification(notification)
-      expect(mail.from).to eq ['postmaster@localhost']
+      expect(mail.from).to eq [blackhole_email]
     end
 
-    it "sets reply_to headers" do
+    it "sets mail headers" do
       mail = NotificationMailer.response_notification(notification)
-      expected_reply_to = "#{AlaveteliConfiguration.contact_name} " \
-                          "<#{AlaveteliConfiguration.contact_email}>"
-      expect(mail.header["Reply-To"].value).to eq expected_reply_to
       expect(mail.header["Return-Path"].value).
         to eq 'do-not-reply-to-this-address@localhost'
     end
@@ -555,14 +549,11 @@ RSpec.describe NotificationMailer do
 
     it "sends the message from the right address" do
       mail = NotificationMailer.embargo_expiring_notification(notification)
-      expect(mail.from).to eq ['postmaster@localhost']
+      expect(mail.from).to eq [blackhole_email]
     end
 
-    it "sets reply_to headers" do
+    it "sets mail headers" do
       mail = NotificationMailer.embargo_expiring_notification(notification)
-      expected_reply_to = "#{AlaveteliConfiguration.contact_name} " \
-                          "<#{AlaveteliConfiguration.contact_email}>"
-      expect(mail.header["Reply-To"].value).to eq expected_reply_to
       expect(mail.header["Return-Path"].value).
         to eq 'do-not-reply-to-this-address@localhost'
     end
@@ -628,14 +619,11 @@ RSpec.describe NotificationMailer do
 
     it 'sends the message from the right address' do
       mail = NotificationMailer.expire_embargo_notification(notification)
-      expect(mail.from).to eq ['postmaster@localhost']
+      expect(mail.from).to eq [blackhole_email]
     end
 
-    it 'sets reply_to headers' do
+    it 'sets mail headers' do
       mail = NotificationMailer.expire_embargo_notification(notification)
-      expected_reply_to = "#{AlaveteliConfiguration.contact_name} " \
-                          "<#{AlaveteliConfiguration.contact_email}>"
-      expect(mail.header['Reply-To'].value).to eq expected_reply_to
       expect(mail.header['Return-Path'].value).
         to eq 'do-not-reply-to-this-address@localhost'
     end
@@ -697,14 +685,11 @@ RSpec.describe NotificationMailer do
 
     it "sends the message from the right address" do
       mail = NotificationMailer.overdue_notification(notification)
-      expect(mail.from).to eq ['postmaster@localhost']
+      expect(mail.from).to eq [blackhole_email]
     end
 
-    it "sets reply_to headers" do
+    it "sets mail headers" do
       mail = NotificationMailer.overdue_notification(notification)
-      expected_reply_to = "#{AlaveteliConfiguration.contact_name} " \
-                          "<#{AlaveteliConfiguration.contact_email}>"
-      expect(mail.header["Reply-To"].value).to eq expected_reply_to
       expect(mail.header["Return-Path"].value).
         to eq 'do-not-reply-to-this-address@localhost'
     end
@@ -768,14 +753,11 @@ RSpec.describe NotificationMailer do
 
     it "sends the message from the right address" do
       mail = NotificationMailer.very_overdue_notification(notification)
-      expect(mail.from).to eq ['postmaster@localhost']
+      expect(mail.from).to eq [blackhole_email]
     end
 
-    it "sets reply_to headers" do
+    it "sets mail headers" do
       mail = NotificationMailer.very_overdue_notification(notification)
-      expected_reply_to = "#{AlaveteliConfiguration.contact_name} " \
-                          "<#{AlaveteliConfiguration.contact_email}>"
-      expect(mail.header["Reply-To"].value).to eq expected_reply_to
       expect(mail.header["Return-Path"].value).
         to eq 'do-not-reply-to-this-address@localhost'
     end